refactor: 메뉴 루트 어그리거트를 패키징하라
