+++ src/main/java/kitchenpos/menus/menu/domain/Menu.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.domain;
+package kitchenpos.menus.menu.domain;
 
 import kitchenpos.menus.menugroup.domain.MenuGroup;
