+++ src/main/java/kitchenpos/menus/menu/domain/MenuRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.domain;
+package kitchenpos.menus.menu.domain;
 
 import java.util.List;
 import java.util.Optional;
