+++ src/test/java/kitchenpos/menus/application/MenuServiceTest.java
@@ -1,9 +1,10 @@
 package kitchenpos.menus.application;
 
-import kitchenpos.menus.domain.Menu;
+import kitchenpos.menus.menu.domain.Menu;
+import kitchenpos.menus.menu.application.MenuService;
 import kitchenpos.menus.menugroup.domain.MenuGroupRepository;
-import kitchenpos.menus.domain.MenuProduct;
-import kitchenpos.menus.domain.MenuRepository;
+import kitchenpos.menus.menu.domain.MenuProduct;
+import kitchenpos.menus.menu.domain.MenuRepository;
 import kitchenpos.products.application.FakeProfanity;
 import kitchenpos.products.application.InMemoryProductRepository;
 import kitchenpos.products.domain.Product;
